// lib/types/customJwtPayload.ts
import { JwtPayload } from "jwt-decode";

export interface CustomJwtPayload extends JwtPayload {
  _id: string;
  memberType: string;
  memberStatus: string;
  memberAuthType: string;
  memberPhone: string;
  memberNick: string;
  memberFullName?: string;
  memberImage?: string | null;
  memberAddress?: string;
  memberDesc?: string;
  memberProperties: number;
  memberRank: number;
  memberArticles: number;
  memberPoints: number;
  memberLikes: number;
  memberViews: number;
  memberWarnings: number;
  memberBlocks: number;
  createdAt: Date | null;
  updatedAt: Date | null;
}